Restructure project folders
