+++ alembic.ini
@@ -50,7 +50,7 @@ version_path_separator = os  # default: use os.pathsep
 # are written from script.py.mako
 # output_encoding = utf-8
 
-sqlalchemy.url = sqlite:///main.db
+sqlalchemy.url = sqlite:///database/main.db
 
 
 [post_write_hooks]
